YQL-19747 Rearrange modules
