+++ yql/essentials/sql/v1/complete/name/service/fallback/name_service.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <yql/essentials/sql/v1/complete/name/name_service.h>
+#include <yql/essentials/sql/v1/complete/name/service/name_service.h>
 
 namespace NSQLComplete {
