+++ yql/essentials/sql/v1/complete/name/service/fallback/ya.make
