+++ yql/essentials/sql/v1/complete/name/service/static/ranking.cpp
@@ -2,7 +2,7 @@
 
 #include "frequency.h"
 
-#include <yql/essentials/sql/v1/complete/name/name_service.h>
+#include <yql/essentials/sql/v1/complete/name/service/name_service.h>
 
 #include <util/charset/utf8.h>
