+++ yql/essentials/sql/v1/complete/name/service/static/ut/ya.make
@@ -1,0 +1,7 @@
+UNITTEST_FOR(yql/essentials/sql/v1/complete/name/service/static)
+
+SRCS(
+    frequency_ut.cpp
+)
+
+END()
