+++ yql/essentials/sql/v1/complete/name/service/ya.make
@@ -1,0 +1,8 @@
+LIBRARY()
+
+END()
+
+RECURSE(
+    fallback
+    static
+)
