+++ yql/essentials/sql/v1/complete/name/static/frequency.h
@@ -1,20 +1,3 @@
 #pragma once
 
-#include <util/generic/string.h>
-#include <util/generic/hash.h>
-
-namespace NSQLComplete {
-
-    struct TFrequencyData {
-        THashMap<TString, size_t> Keywords;
-        THashMap<TString, size_t> Pragmas;
-        THashMap<TString, size_t> Types;
-        THashMap<TString, size_t> Functions;
-        THashMap<TString, size_t> Hints;
-    };
-
-    TFrequencyData ParseJsonFrequencyData(const TStringBuf text);
-
-    TFrequencyData LoadFrequencyData();
-
-} // namespace NSQLComplete
+#include <yql/essentials/sql/v1/complete/name/service/static/ranking.h>
