--- yql/essentials/sql/v1/complete/name/static/ut/ya.make
@@ -1,7 +1,0 @@
-UNITTEST_FOR(yql/essentials/sql/v1/complete/name/static)
-
-SRCS(
-    frequency_ut.cpp
-)
-
-END()
