+++ yql/essentials/sql/v1/complete/name/ya.make
@@ -2,12 +2,13 @@ LIBRARY()
 
 PEERDIR(
     yql/essentials/sql/v1/complete/core
+    yql/essentials/sql/v1/complete/name/service
 )
 
 END()
 
 RECURSE(
-    fallback
     schema
+    service
     static
 )
