+++ yql/essentials/sql/v1/complete/ut/ya.make
@@ -5,7 +5,7 @@ SRCS(
 )
 
 PEERDIR(
-    yql/essentials/sql/v1/complete/name/fallback
+    yql/essentials/sql/v1/complete/name/service/fallback
     yql/essentials/sql/v1/lexer/antlr4_pure
     yql/essentials/sql/v1/lexer/antlr4_pure_ansi
 )
